Fix the APY formula examples derived from it
The APY and all examples derived from it were missing the removal of the initially included principal stand-in.
I.e. just as the principal stand-in is added to the reward yield at the beginning of the formula ( 1 + reward yield ) it also has to be removed at the end ( - 1 ) so that we only get the yield/gain, not the total balance.

Correct formula:
APY = (1 + reward yield) ^ (3 x 365) - 1

To calculate the reward yield from an APY, we have to do the reverse, i.e. *add* the principal stand-in to the APY:
Reward yield = (1 + APY) ^ (1 / 3 x 365) - 1